+++ shared/src/test/java/lemon/shared/test/message/MsgRepositoryTest.java
@@ -183,7 +183,7 @@ public class MsgRepositoryTest extends BaseMmtTest {
 	}
 
 	private void prepareMsg(Message msg) {
-		msg.setCreateTime(System.currentTimeMillis() / 1000);
+		msg.setCreateTime((int) (System.currentTimeMillis() / 1000));
 		msg.setCust_id(CUST_ID);
 		msg.setFromUserName("fromUserName");
 		msg.setId(IdWorkerManager.getIdWorker(Message.class).getId());
